Add files via upload
This model performs clustering on the Iris dataset using the K-means algorithm and the Elbow Method to determine the optimal number of clusters. The Iris data is loaded using the sciktit-learn library and normalized using StandardScaler. The K-means model is then trained for different numbers of clusters (from 1 to 10), and the inertia for each is calculated and stored in a list. The inertia values are used to plot the Elbow Method graph with the matplotlib library, where the X-axis represents the number of clusters and the Y-axis represents inertia. The result is a graph that helps identify the optimal number of clusters at the point where the graph bends sharply.